+++ src/shared/components/PluginInfo/index.tsx
