+++ src/shared/containers/ResourcePlugins.tsx
@@ -3,7 +3,7 @@ import { Resource } from '@bbp/nexus-sdk';
 import { Collapse } from 'antd';
 import { useNexusContext } from '@bbp/react-nexus';
 import { NexusPlugin } from '../containers/NexusPlugin';
-import PluginInfo from './PluginInfo';
+import PluginInfo from '../components/PluginInfo';
 import { matchPlugins, pluginsMap } from '../utils';
 import usePlugins from '../hooks/usePlugins';
